requirement.txt
BeautifulSoup==3.2.1
Django==1.11.3
SQLAlchemy==1.1.12
argparse==1.2.1
django-extensions==1.8.1
djangorestframework==3.6.3
et-xmlfile==1.0.1
gunicorn==19.7.1
jdcal==1.3
numpy==1.13.1
openpyxl==2.4.8
pandas==0.20.3
psycopg2==2.7.2
pycrypto==2.6.1
python-dateutil==2.6.1
pytz==2017.2
six==1.10.0
wsgiref==0.1.2
xlrd==1.0.0
